+++ src/FormattingOptions.ts
@@ -1,3 +1,6 @@
+import { Formatter } from './Formatter';
+import { TokenKind } from 'brighterscript';
+
 /**
  * A set of formatting options used to determine how the file should be formatted.
  */
@@ -50,7 +53,11 @@ export interface FormattingOptions {
     typeCaseOverride?: { [id: string]: FormattingOptions['keywordCase'] };
     /**
      * If true (the default), all whitespace between items is reduced to exactly 1 space character,
-     * and certain keywords and operators are padded with whitespace (i.e. `1+1` becomes `1 + 1`)
+     * and certain keywords and operators are padded with whitespace (i.e. `1+1` becomes `1 + 1`).
+     * This is a catchall property that will also disable the following rules:
+     * - insertSpaceBeforeFunctionParenthesis
+     * - insertSpaceBetweenEmptyCurlyBraces
+     * - insertSpaceAfterOpeningAndBeforeClosingNonemptyBraces
      */
     formatInteriorWhitespace?: boolean;
     /**
@@ -66,10 +73,92 @@ export interface FormattingOptions {
      */
     insertSpaceBetweenEmptyCurlyBraces?: boolean;
     /**
+     * If true, ensure exactly 1 space after leading and before trailing curly braces
+     * If false, REMOVE all whitespace after leading and before trailing curly braces (excluding beginning-of-line indentation spacing)
+     * @default true
+     */
+    insertSpaceAfterOpeningAndBeforeClosingNonemptyBraces?: boolean;
+    /**
      * Forces all single-line comments to use the same style.
      * If 'singlequote' or falsey, all comments are preceeded by a single quote. This is the default.
      * If 'rem', all comments are preceeded by `rem`
      * If 'original', the comment type is unchanged
      */
     formatSingleLineCommentType?: 'singlequote' | 'rem' | 'original';
+}
+
+
+export function normalizeOptions(options: FormattingOptions | undefined = {}) {
+    let fullOptions: FormattingOptions = {
+        indentStyle: 'spaces',
+        indentSpaceCount: Formatter.DEFAULT_INDENT_SPACE_COUNT,
+        formatIndent: true,
+        keywordCase: 'lower',
+        compositeKeywords: 'split',
+        removeTrailingWhiteSpace: true,
+        keywordCaseOverride: {},
+        formatInteriorWhitespace: true,
+        insertSpaceBeforeFunctionParenthesis: false,
+        insertSpaceBetweenEmptyCurlyBraces: false,
+        insertSpaceAfterOpeningAndBeforeClosingNonemptyBraces: true,
+
+        //override defaults with the provided values
+        ...options
+    };
+
+    if (!fullOptions.typeCase) {
+        fullOptions.typeCase = fullOptions.keywordCase as any;
+    }
+
+    fullOptions.keywordCaseOverride = normalizeKeywordCaseOverride(fullOptions.keywordCaseOverride);
+    fullOptions.typeCaseOverride = normalizeKeywordCaseOverride(fullOptions.typeCaseOverride);
+
+    return fullOptions;
+}
+
+export function normalizeKeywordCaseOverride(obj: FormattingOptions['keywordCaseOverride']) {
+    let result = {};
+
+    //quit now if the object is not iterable
+    if (!obj) {
+        return result;
+    }
+
+    for (let key in obj) {
+        let value = obj[key]
+            ? obj[key]!.toLowerCase()
+            : 'disabled';
+
+        if (value === 'original') {
+            value = 'disabled';
+        }
+
+        key = key
+            //remove any whitespace
+            .replace(/\s+/gi, '')
+            //force key to lower case
+            .toLowerCase();
+
+        //replace some of the hash tokens with their corresponding TokenKind
+        if (key === '#const') {
+            key = TokenKind.HashConst.toLowerCase();
+
+        } else if (key === '#else') {
+            key = TokenKind.HashElse.toLowerCase();
+
+        } else if (key === '#elseif') {
+            key = TokenKind.HashElseIf.toLowerCase();
+
+        } else if (key === '#endif') {
+            key = TokenKind.HashEndIf.toLowerCase();
+
+        } else if (key === '#error') {
+            key = TokenKind.HashError.toLowerCase();
+
+        } else if (key === '#if') {
+            key = TokenKind.HashIf.toLowerCase();
+        }
+        result[key] = value;
+    }
+    return result;
 }